refactor(utils): lift resolveReorderAnchor to a shared reorder util
Generic over any {id}[] list, and the dashboard deck reorder needs the
same anchor-resolution logic the card editor already has.
